// frontend/reporting/src/app/models/usage-details-info.model.ts
export class UsageDetailsInfo {
  type1 = 'usage';
  type2: string;
  type3: string;

  code1: string;
  code2: string;
  code3: string;

  status1: string;
  status2: string;
  status3: string;

  text1: string;
  text2: string;
  text3: string;

  clob1: string;
  clob2: string;
  clob3: string;
}